// frontend/src/openapi/models/ModelSwapPrompt.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ModelSwapPrompt
 */
export interface ModelSwapPrompt {
    /**
     *
     * @type {string}
     * @memberof ModelSwapPrompt
     */
    prompt: string;
    /**
     *
     * @type {boolean}
     * @memberof ModelSwapPrompt
     */
    pin: boolean;
    /**
     *
     * @type {string}
     * @memberof ModelSwapPrompt
     */
    capabilityId: string;
    /**
     *
     * @type {string}
     * @memberof ModelSwapPrompt
     */
    capabilityName: string;
    /**
     *
     * @type {string}
     * @memberof ModelSwapPrompt
     */
    oldModel: string;
    /**
     *
     * @type {string}
     * @memberof ModelSwapPrompt
     */
    newModel: string;
}

/**
 * Check if a given object implements the ModelSwapPrompt interface.
 */
export function instanceOfModelSwapPrompt(value: object): value is ModelSwapPrompt {
    if (!('prompt' in value) || value['prompt'] === undefined) return false;
    if (!('pin' in value) || value['pin'] === undefined) return false;
    if (!('capabilityId' in value) || value['capabilityId'] === undefined) return false;
    if (!('capabilityName' in value) || value['capabilityName'] === undefined) return false;
    if (!('oldModel' in value) || value['oldModel'] === undefined) return false;
    if (!('newModel' in value) || value['newModel'] === undefined) return false;
    return true;
}

export function ModelSwapPromptFromJSON(json: any): ModelSwapPrompt {
    return ModelSwapPromptFromJSONTyped(json, false);
}

export function ModelSwapPromptFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelSwapPrompt {
    if (json == null) {
        return json;
    }
    return {

        'prompt': json['prompt'],
        'pin': json['pin'],
        'capabilityId': json['capability_id'],
        'capabilityName': json['capability_name'],
        'oldModel': json['old_model'],
        'newModel': json['new_model'],
    };
}

export function ModelSwapPromptToJSON(json: any): ModelSwapPrompt {
    return ModelSwapPromptToJSONTyped(json, false);
}

export function ModelSwapPromptToJSONTyped(value?: ModelSwapPrompt | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'prompt': value['prompt'],
        'pin': value['pin'],
        'capability_id': value['capabilityId'],
        'capability_name': value['capabilityName'],
        'old_model': value['oldModel'],
        'new_model': value['newModel'],
    };
}